moat: 11 mechanisms, hole-first frontier, survivor registry -- and CI has never been green
THE MOAT HUNT IS NOW CONTINUOUS AND CONVERGENT.

  ELEVEN MECHANISMS, NOT SEVEN. libs/hypmax/moat_features.py builds the four specs the
  asymmetry ledger ranked first and that had sat as proposals: hidden_liquidity (traded size
  minus the displayed size that vanished = inferred iceberg residual), queue_depth_at_touch
  (resting size in units of the volume that then traded -- sizes adverse selection directly),
  book_pressure_vs_funding (FUSE class; returns EMPTY when the venue publishes no funding,
  because defaulting the missing leg to zero turns it into plain imbalance wearing a better
  name), spread_regime_shift (trailing median ending at the PREVIOUS snapshot). Every one is
  RECONSTRUCT or FUSE class: computed from timestamps only this desk holds.

  HOLE-FIRST SCHEDULING. The screen took files[-200:] every run -- it re-screened the newest
  slice forever and never once looked at the older archive, the part a competitor most
  conclusively cannot obtain because it cannot be backfilled at any price. Coverage of the
  (venue, symbol, day) x mechanism grid now persists and the budget goes to the cells owing the
  most mechanisms: the same scheduler that made mine_moat converge, over the same grid, so
  mined-but-unscreened is a visible hole rather than an accounting mismatch.

  SURVIVORS PERSIST, WITH THEIR MISSES. Romano-Wolf controls family-wise error WITHIN a run and
  nothing controls it ACROSS runs, so screening the archive repeatedly returns false survivors
  at the nominal rate by construction. The registry keeps every promoted triple with the
  denominator -- times screened, hit rate, IC sign stability -- and only >=2 survivals on >=2
  independent cells counts as PERSISTENT.

  quant-moat-screen.service runs it 24/7, nice'd BELOW the miner which is below the recorders:
  lost tape is unbuyable, a delayed pass costs minutes.

TWO REAL DEFECTS FOUND WHILE WIRING IT.

  A FULL-SAMPLE Z-SCORE IN MY OWN NEW FEATURE. book_pressure_vs_funding standardised funding
  against its complete mean and std, so the z at hour one knew what hour twenty did. Replaced
  with an expanding z. It survived review only because the synthetic tape publishes no funding,
  so the truncation test skipped it -- an untested branch is not a tested one. The test tape now
  carries funding and the causality check covers all four features.

  A GAP IS NOT A HORIZON. Subsampling by stride assumes every step is stride x 15s; across a
  recorder outage -- or, now that scheduling is hole-first, across two non-adjacent days -- one
  "60-second period" can be twelve hours, carrying twelve hours of return into a sample whose
  every other point carries a minute. Periods are now masked to their nominal span.

AND THE FINDING THAT MATTERS MORE THAN ANY OF THAT: CI HAS NEVER BEEN GREEN.

  Every commit in this repo's visible history fails `ruff check .` -- 299 errors at the first
  commit, 448 at HEAD. Because the lint step fails, the workflow SKIPS mypy and pytest. The
  quality gate has never once run the type checker or the test suite on any commit, and I have
  been reporting "tests passing" from local runs without ever looking at it.

  All three gates are now green: ruff clean, mypy --strict clean across 412 source files, 2404
  tests passing. Fixes were mechanical (ruff format, then parameterising bare generics and
  annotating untyped defs) except where they were not:
    - moat_audit dropped the parsed bid/ask QUANTITY, which is the validation that rejects an
      unparseable row before it reaches the spread maths -- kept, renamed.
    - screen_smart_dumb re-searched every result row inside its own print loop to find a row it
      had appended two lines above.
    - run_trade_forensics compared three heterogeneous dict values in one condition.
    - Seven one-shot scratch scripts in the repo root (par.py, wire.py, esc.py, fixsweep.py,
      _audit_gate_probe*.py, _audit_rows.pkl) referenced by nothing, already executed, carrying
      22 of the errors: deleted.

+++ libs/research/mine_conversion.py
@@ -847,7 +847,7 @@ def _rate_se(rate: float, n: int) -> float:
     """Binomial standard error, floored so n=1 cannot report perfect precision."""
     if n <= 0:
         return 1.0
-    return max((rate * (1.0 - rate) / n) ** 0.5, 1.0 / n)
+    return float(max((rate * (1.0 - rate) / n) ** 0.5, 1.0 / n))
 
 
 def law_effectiveness(
